fix(symbol): change how symbol scope is checked
I need to change where the scope changes from global to local. The
rule is that when I hit a LPAREN token when checking if the rule is
decl_or_func, I'm entering a local scope. If I hit a comma, I add the ID
lexemes to the global symbol table.