Assignment 2 milestone 2 (#5)
* fix: id_list not detecting duplicates

* fix: undeclared functions not throwing error

* fix: variable called as function did not fail

* fix: formal symbols not added to table

* fix: not checking scope of symbols

* chore: move grammar rules to separate directory

* docs: update tasks for assignment 2 milestone 2

* feat: add g2 grammar rules spec

* fix: missing nonterminals in grammar file

* chore: add gff to gitignore

* feat: add g2-first-follow

* feat: add g2 tokens

* wip!: adding g2 grammar rule parsing

* feat!: add token peek

* fix: lexeme was being updated by the peek

* refactor: change missed old lexemes

* fix: missing else token

* refactor: add feature directories

* build: Makefile now automatically finds files

- Compiles differently on gradescope vs locally based on structure

* refactor: add better grammar_rule handling

* fix: fixed lookup seg faulting

* chore(gitignore): ignore submission directory

* refactor(grammar): use left-factoring

An issue with the current grammar rules are that I didn't implement
left-factoring. As such, whenever I encounter a grammar rule where I
cannot determine the next rule with 1-token lookahead, I have to do some
weird solution. Left-factoring will fix this.

- added decl_or_func grammar rule
- changed prog grammar rule to use decl_or_func

* chore: remove unused parser_service

Moved to using a refined grammar rules instead of the prior mess that
was parser_service

* chore: remove unused files

* fix(symbol): change how symbol scope is checked

I need to change where the scope changes from global to local. The
rule is that when I hit a LPAREN token when checking if the rule is
decl_or_func, I'm entering a local scope. If I hit a comma, I add the ID
lexemes to the global symbol table.

* fix(grammar): lookup in local doesn't find global

- Issue looks to be that lookup from within a local scope doesn't find
the symbol in the global table and throws an error

- Found the cause. decl_or_func grammar rule wasn't added ID to the
symbol table if the SEMI rule was matched. Moved ID symbol adding to
outside the rules since all 3 needed them anyways.

* fix(sem): fn_call no error on wrong num. of args

Issue: fn_call isn't throwing an error when it's called with the wrong
number of arguments. If f(int x) is defined, f(x, x, x) should throw an
error as undefined.

* test: testing parser

* test: testing semantics